% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{all_nominal}
\alias{all_nominal}
\title{Find all nominal variables.}
\usage{
all_nominal()
}
\description{
Currently we only consider "string" type as nominal.
}
\seealso{
Other Selectors: 
\code{\link{all_numeric}()},
\code{\link{has_type}()}
}
\concept{Selectors}
