% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{all_numeric}
\alias{all_numeric}
\title{Speciy all numeric variables.}
\usage{
all_numeric()
}
\description{
Find all the variables with the following types:
"float16", "float32", "float64", "int16", "int32", "int64",
"half", "double".
}
\seealso{
Other Selectors: 
\code{\link{all_nominal}()},
\code{\link{has_type}()}
}
\concept{Selectors}
