% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{as_array_iterator}
\alias{as_array_iterator}
\title{Convert tf_dataset to an iterator that yields R arrays.}
\usage{
as_array_iterator(dataset)
}
\arguments{
\item{dataset}{A tensorflow dataset}
}
\value{
An iterable. Use \code{\link[=iterate]{iterate()}} or \code{\link[=iter_next]{iter_next()}} to access values from the iterator.
}
\description{
Convert tf_dataset to an iterator that yields R arrays.
}
