% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{as_tf_dataset}
\alias{as_tf_dataset}
\title{Add the tf_dataset class to a dataset}
\usage{
as_tf_dataset(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
A dataset with class "tf_dataset"
}
\description{
Calling this function on a dataset adds the "tf_dataset" class to the dataset
object. All datasets returned by functions in the \pkg{tfdatasets} package
call this function on the dataset before returning it.
}
\keyword{internal}
