% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_line_dataset.R
\name{csv_dataset}
\alias{csv_dataset}
\title{Create a dataset from a text file with comma separated values}
\usage{
csv_dataset(filenames, compression_type = NULL, col_names = TRUE,
  record_defaults = NULL, field_delim = ",", skip = 0,
  num_threads = NULL, output_buffer_size = NULL)
}
\arguments{
\item{filenames}{Characater vector containing one or more filenames.}

\item{compression_type}{A string, one of: \code{"auto"} (determine based on file
extension), \code{""} (no compression), \code{"ZLIB"}, or \code{"GZIP"}. For
\code{"auto"}, GZIP will be automatically selected if any of
the \code{filenames} have a .gz extension and ZLIB will be automatically
selected if any of the \code{filenames} have a .zlib extension (otherwise
no compression will be used).}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector of column
names.

If \code{TRUE}, the first row of the input will be used as the column names, and
will not be included in dataset. If \code{FALSE}, column names will be generated
automatically: X1, X2, X3 etc.

If \code{col_names} is a character vector, the values will be used as the names
of the columns, and the first row of the input will be read into the first
row of the datset.}

\item{record_defaults}{List of default values for records. Default values must
be of type integer, numeric, or character. Used both to indicate the type
of each field as well as to provide defaults for missing values.}

\item{field_delim}{An optional string. Defaults to ",". char delimiter to
separate fields in a record.}

\item{skip}{Number of lines to skip before reading data. Note that if
\code{col_names} is explicitly provided and there are column names witin the CSV
file then \code{skip} should be set to 1 to ensure that the column names are
bypassed.}

\item{num_threads}{(Optional) An integer representing the number of threads
to use for processing elements in parallel. If not specified, elements will
be processed sequentially without buffering.}

\item{output_buffer_size}{(Optional) An integer representing the maximum
number of processed elements that will be buffered when processing in
parallel.}
}
\description{
Create a dataset from a text file with comma separated values
}
\section{Column Names}{


Column names are not an intrinsic property of TensorFlow datasets, however
they are supported in this interface to facilitate specifying features and
response variables when creating input functions and generators that draw
from the dataset.
}

