% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_cache}
\alias{dataset_cache}
\title{Caches the elements in this dataset.}
\usage{
dataset_cache(dataset, filename = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{filename}{String with the name of a directory on the filesystem to use
for caching tensors in this Dataset. If a filename is not provided, the
dataset will be cached in memory.}
}
\value{
A dataset
}
\description{
Caches the elements in this dataset.
}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_map}}, \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle}},
  \code{\link{dataset_take}}, \code{\link{dataset_unbatch}}
}
