% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_concatenate}
\alias{dataset_concatenate}
\title{Creates a dataset by concatenating given dataset with this dataset.}
\usage{
dataset_concatenate(dataset, other)
}
\arguments{
\item{dataset}{A dataset}

\item{other}{Dataset to be concatenated}
}
\value{
A dataset
}
\description{
Creates a dataset by concatenating given dataset with this dataset.
}
\note{
Input dataset and dataset to be concatenated should have same nested
structures and output types.
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_map}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_prefetch}()},
\code{\link{dataset_repeat}()},
\code{\link{dataset_shuffle_and_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_skip}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
