% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_enumerate}
\alias{dataset_enumerate}
\title{Enumerate the elements of this dataset.}
\usage{
dataset_enumerate(dataset, start = 0)
}
\arguments{
\item{dataset}{A dataset}

\item{start}{A integer, representing the start value for enumeration.}
}
\value{
A dataset
}
\description{
Adds a counter to an iterable. So for each element in the dataset, a tuple is produced with (counter, element)
}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_cache}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_decode_csv}},
  \code{\link{dataset_map}}, \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle}},
  \code{\link{dataset_skip}}, \code{\link{dataset_take}},
  \code{\link{dataset_unbatch}}
}
