% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_flat_map}
\alias{dataset_flat_map}
\title{Maps map_func across this dataset and flattens the result.}
\usage{
dataset_flat_map(dataset, map_func)
}
\arguments{
\item{dataset}{A dataset}

\item{map_func}{A function mapping a nested structure of tensors (having
shapes and types defined by \code{\link[=output_shapes]{output_shapes()}} and \code{\link[=output_types]{output_types()}} to a
dataset.}
}
\value{
A dataset
}
\description{
Maps map_func across this dataset and flattens the result.
}
