% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_tensors.R
\name{dataset_from_tensor_slices}
\alias{dataset_from_tensor_slices}
\title{Creates a dataset whose elements are slices of the given tensors.}
\usage{
dataset_from_tensor_slices(tensors)
}
\arguments{
\item{tensors}{A nested structure of tensors, each having the same size in
the 0th dimension.}
}
\value{
A dataset.
}
\description{
Creates a dataset whose elements are slices of the given tensors.
}
\seealso{
Other tensor datasets: \code{\link{dataset_from_sparse_tensor_slices}},
  \code{\link{dataset_from_tensors}}
}
