% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_ignore_errors}
\alias{dataset_ignore_errors}
\title{Creates a Dataset from this one and silently ignores any errors.}
\usage{
dataset_ignore_errors(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
A dataset
}
\description{
Use this transformation to produce a dataset that contains the same elements
as the input, but silently drops any elements that caused an error.
}
\examples{
\dontrun{

# tf$check_numerics(1 / 0) will raise an error which is ignored

dataset <- tensor_slices_dataset(list(1, 2, 0, 4)) \%>\%
  dataset_map(function(x) {
    tf$check_numerics(1 / x, "error")
  }) \%>\%
  dataset_ignore_errors()
}

}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_cache}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_decode_csv}},
  \code{\link{dataset_enumerate}},
  \code{\link{dataset_map}}, \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle}},
  \code{\link{dataset_skip}}, \code{\link{dataset_take}},
  \code{\link{dataset_unbatch}}
}
