% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_map}
\alias{dataset_map}
\title{Map a function across a dataset.}
\usage{
dataset_map(dataset, map_func, num_parallel_calls = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{map_func}{A function mapping a nested structure of tensors (having
shapes and types defined by \code{\link[=output_shapes]{output_shapes()}} and \code{\link[=output_types]{output_types()}} to
another nested structure of tensors. It also supports \code{purrr} style
lambda functions powered by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{num_parallel_calls}{(Optional) An integer, representing the
number of elements to process in parallel If not specified, elements will
be processed sequentially.}
}
\value{
A dataset
}
\description{
Map a function across a dataset.
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_concatenate}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_prefetch}()},
\code{\link{dataset_repeat}()},
\code{\link{dataset_shuffle_and_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_skip}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
