% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_prefetch}
\alias{dataset_prefetch}
\title{Creates a Dataset that prefetches elements from this dataset.}
\usage{
dataset_prefetch(dataset, buffer_size = tf$data$AUTOTUNE)
}
\arguments{
\item{dataset}{A dataset}

\item{buffer_size}{An integer, representing the maximum number elements that
will be buffered when prefetching.}
}
\value{
A dataset
}
\description{
Creates a Dataset that prefetches elements from this dataset.
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_concatenate}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_map}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_reduce}()},
\code{\link{dataset_repeat}()},
\code{\link{dataset_shuffle_and_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_skip}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
