% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_repeat}
\alias{dataset_repeat}
\title{Repeats a dataset count times.}
\usage{
dataset_repeat(dataset, count = NULL)
}
\arguments{
\item{count}{(Optional.) An integer value representing the number of times
the elements of this dataset should be repeated. The default behavior (if
\code{count} is \code{NULL} or \code{-1}) is for the elements to be repeated indefinitely.}
}
\value{
A dataset
}
\description{
Repeats a dataset count times.
}
