% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_shard}
\alias{dataset_shard}
\title{Creates a dataset that includes only 1 / num_shards of this dataset.}
\usage{
dataset_shard(dataset, num_shards, index)
}
\arguments{
\item{dataset}{A dataset}

\item{num_shards}{A integer representing the number of shards operating in
parallel.}

\item{index}{A integer, representing the worker index.}
}
\value{
A dataset
}
\description{
This dataset operator is very useful when running distributed training, as it
allows each worker to read a unique subset.
}
\concept{Dataset methods}
