% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_shuffle_and_repeat}
\alias{dataset_shuffle_and_repeat}
\title{Shuffles and repeats a dataset returning a new permutation for each epoch.}
\usage{
dataset_shuffle_and_repeat(dataset, buffer_size, count = NULL, seed = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{buffer_size}{An integer, representing the number of elements from this
dataset from which the new dataset will sample.}

\item{count}{(Optional.) An integer value representing the number of times
the elements of this dataset should be repeated. The default behavior (if
\code{count} is \code{NULL} or \code{-1}) is for the elements to be repeated indefinitely.}

\item{seed}{(Optional) An integer, representing the random seed that will be
used to create the distribution.}
}
\description{
Shuffles and repeats a dataset returning a new permutation for each epoch.
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_concatenate}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_map}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_prefetch}()},
\code{\link{dataset_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_skip}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
