% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_skip}
\alias{dataset_skip}
\title{Creates a dataset that skips count elements from this dataset}
\usage{
dataset_skip(dataset, count)
}
\arguments{
\item{dataset}{A dataset}

\item{count}{An integer, representing the number of elements of this dataset
that should be skipped to form the new dataset. If count is greater than
the size of this dataset, the new dataset will contain no elements. If
count is -1, skips the entire dataset.}
}
\value{
A dataset
}
\description{
Creates a dataset that skips count elements from this dataset
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_concatenate}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_map}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_prefetch}()},
\code{\link{dataset_repeat}()},
\code{\link{dataset_shuffle_and_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
