% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_tensors.R
\name{dataset_sparse_tensor_slices}
\alias{dataset_sparse_tensor_slices}
\title{Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.}
\usage{
dataset_sparse_tensor_slices(sparse_tensor)
}
\arguments{
\item{sparse_tensor}{A \code{tf$SparseTensor}.}
}
\value{
A dataset of rank-(N-1) sparse tensors.
}
\description{
Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.
}
\seealso{
Other tensor datasets: \code{\link{dataset_tensor_slices}},
  \code{\link{dataset_tensors}}
}
