% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_take}
\alias{dataset_take}
\title{Creates a dataset with at most count elements from this dataset}
\usage{
dataset_take(dataset, count)
}
\arguments{
\item{dataset}{A dataset}

\item{count}{Integer representing the number of elements of this dataset that
should be taken to form the new dataset. If \code{count} is -1, or if \code{count} is
greater than the size of this dataset, the new dataset will contain all
elements of this dataset.}
}
\value{
A dataset
}
\description{
Creates a dataset with at most count elements from this dataset
}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_cache}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_decode_delim}},
  \code{\link{dataset_filter}},
  \code{\link{dataset_interleave}},
  \code{\link{dataset_map_and_batch}},
  \code{\link{dataset_map}},
  \code{\link{dataset_padded_batch}},
  \code{\link{dataset_prefetch_to_device}},
  \code{\link{dataset_prefetch}},
  \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle_and_repeat}},
  \code{\link{dataset_shuffle}}, \code{\link{dataset_skip}}
}
