% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_tensors.R
\name{dataset_tensors}
\alias{dataset_tensors}
\title{Creates a dataset with a single element, comprising the given tensors.}
\usage{
dataset_tensors(tensors)
}
\arguments{
\item{tensors}{A nested structure of tensors.}
}
\value{
A dataset.
}
\description{
Creates a dataset with a single element, comprising the given tensors.
}
\seealso{
Other tensor datasets: \code{\link{dataset_sparse_tensor_slices}},
  \code{\link{dataset_tensor_slices}}
}
