% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_unbatch}
\alias{dataset_unbatch}
\title{Splits elements of this dataset into sequences of consecutive elements.}
\usage{
dataset_unbatch(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
A dataset
}
\description{
For example, if elements of this dataset are shaped \code{[B, a0, a1, ...]}, where B
may vary from element to element, then for each element in this dataset, the
unbatched dataset will contain B consecutive elements of shape \code{[a0, a1, ...]}.
}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_cache}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_enumerate}},
  \code{\link{dataset_map}}, \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle}},
  \code{\link{dataset_skip}}, \code{\link{dataset_take}}
}
