% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{dense_features}
\alias{dense_features}
\title{Dense Features}
\usage{
dense_features(spec)
}
\arguments{
\item{spec}{A feature specification created with \code{\link[=feature_spec]{feature_spec()}}.}
}
\value{
A list of feature columns.
}
\description{
Retrives the Dense Features from a spec.
}
