% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{fit.FeatureSpec}
\alias{fit.FeatureSpec}
\title{Fits a feature specification.}
\usage{
\method{fit}{FeatureSpec}(spec, dataset = NULL, ...)
}
\arguments{
\item{spec}{A feature specification created with \code{\link[=feature_spec]{feature_spec()}}.}

\item{dataset}{(Optional) A TensorFlow dataset. If \code{NULL} it will use
the dataset provided when initilializing the \code{feature_spec}.}

\item{...}{(unused)}
}
\value{
a fitted \code{FeatureSpec} object.
}
\description{
This function will \code{fit} the specification. Depending
on the steps added to the specification it will compute
for example, the levels of categorical features, normalization
constants, etc.
}
\examples{
\dontrun{
library(tfdatasets)
data(hearts)
hearts <- tensor_slices_dataset(hearts) \%>\% dataset_batch(32)

# use the formula interface
spec <- feature_spec(hearts, target ~ age) \%>\%
  step_numeric_column(age)

spec_fit <- fit(spec)
spec_fit
}
}
\seealso{
\itemize{
\item \code{\link[=feature_spec]{feature_spec()}} to initialize the feature specification.
\item \code{\link[=dataset_use_spec]{dataset_use_spec()}} to create a tensorflow dataset prepared to modeling.
\item \link{steps} to a list of all implemented steps.
}

Other Feature Spec Functions: \code{\link{dataset_use_spec}},
  \code{\link{feature_spec}}, \code{\link{steps}}
}
\concept{Feature Spec Functions}
