% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_length_dataset.R
\name{fixed_length_record_dataset}
\alias{fixed_length_record_dataset}
\title{A dataset of fixed-length records from one or more binary files.}
\usage{
fixed_length_record_dataset(
  filenames,
  record_bytes,
  header_bytes = NULL,
  footer_bytes = NULL,
  buffer_size = NULL
)
}
\arguments{
\item{filenames}{A string tensor containing one or more filenames.}

\item{record_bytes}{An integer representing the number of bytes in each
record.}

\item{header_bytes}{(Optional) An integer scalar representing the number of
bytes to skip at the start of a file.}

\item{footer_bytes}{(Optional) A integer scalar representing the number of
bytes to ignore at the end of a file.}

\item{buffer_size}{(Optional) A integer scalar representing the number of
bytes to buffer when reading.}
}
\value{
A dataset
}
\description{
A dataset of fixed-length records from one or more binary files.
}
