% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{for_each_batch}
\alias{for_each_batch}
\title{Iterate over a dataset, calling a function for each batch}
\usage{
for_each_batch(dataset, func)
}
\arguments{
\item{dataset}{A dataset}

\item{func}{Function to call with batch tensor

Traverse a dataset by calling a function that accepts the next
batch as a parameter. Note that when running in default (non-eager)
mode, the tensor must actually be evaluated (e.g. passed to
a TensorFlow function that uses it or run explicitly) in order
to advance to the next batch.}
}
\description{
Iterate over a dataset, calling a function for each batch
}
