% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generator_from_dataset}
\alias{generator_from_dataset}
\title{Construct a Keras generator function from a dataset}
\usage{
generator_from_dataset(dataset, features, response, to_categorical = NULL,
  session = "default")
}
\arguments{
\item{dataset}{A dataset}

\item{features}{The names of feature variables to be used.}

\item{response}{The name of the response variable.}

\item{session}{TensorFlow session to evaluate values within. If
"default", then the current Keras TensorFlow session will be
used (if available) and failing that the default TensorFlow
session as given by \code{tf$get_default_session()}.}

\item{filter}{Optional filter function which}
}
\value{
A generator function suitable for use with the keras
\link[keras:fit_generator]{fit_generator},
\href{keras::evaluate_generator}{evaluate_generator}, and
\href{keras::predict_generator}{predict_generator} functions.
}
\description{
Construct a Keras generator function from a dataset
}
\details{
Creating a generator from a dataset requires that the dataset
consist of a set of named output tensors (e.g. like the dataset produced by
the \code{\link[=csv_dataset]{csv_dataset()}} function).
}
