% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{has_type}
\alias{has_type}
\title{Identify the type of the variable.}
\usage{
has_type(match = "float32")
}
\arguments{
\item{match}{A list of types to match.}
}
\description{
Can only be used inside the \link{steps} specifications to find
variables by type.
}
\seealso{
Other Selectors: 
\code{\link{all_nominal}()},
\code{\link{all_numeric}()}
}
\concept{Selectors}
