% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{initializable_iterator}
\alias{initializable_iterator}
\title{Creates an Iterator for enumerating the elements of this dataset.}
\usage{
initializable_iterator(dataset, shared_name = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{shared_name}{(Optional) If non-empty, this iterator will be shared under the
given name across multiple sessions that share the same devices (e.g. when
using a remote server).}
}
\value{
An iterator over the elements of this dataset
}
\description{
The returned iterator will be in an uninitialized state, and you must run the
iterator.initializer operation before using it.
}
\seealso{
Other dataset iterators: \code{\link{iterator_next}},
  \code{\link{one_shot_iterator}}
}
