% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn.tensorflow.python.data.ops.dataset_ops.Dataset}
\alias{input_fn.tensorflow.python.data.ops.dataset_ops.Dataset}
\title{Construct an input function from a dataset}
\usage{
input_fn.tensorflow.python.data.ops.dataset_ops.Dataset(dataset, features,
  response)
}
\arguments{
\item{dataset}{A dataset}

\item{features}{The names of feature variables to be used.}

\item{response}{The name of the response variable.}
}
\value{
An input_fn suitable for use with tfestimators train, evaluate, and
predict methods
}
\description{
Construct an input function from a dataset
}
\details{
Creating an input_fn from a dataset requires that the dataset
consist of a set of named output tensors (e.g. like the dataset
produced by the \code{\link[=csv_dataset]{csv_dataset()}} function).
}
