% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn_from_dataset}
\alias{input_fn_from_dataset}
\title{Construct a tfestimators input function from a dataset}
\usage{
input_fn_from_dataset(dataset, features, response = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{features}{The names of feature variables to be used.}

\item{response}{The name of the response variable.}
}
\value{
An input_fn suitable for use with tfestimators \link[tfestimators:train.tf_estimator]{train}, \link[tfestimators:evaluate.tf_estimator]{evaluate}, and
\link[tfestimators:predict.tf_estimator]{predict} methods
}
\description{
Construct a tfestimators input function from a dataset
}
\details{
Creating an input_fn from a dataset requires that the dataset
consist of a set of named output tensors (e.g. like the dataset
produced by the \code{\link[=csv_dataset]{csv_dataset()}} function).
}
