% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{iterator_from_dataset}
\alias{iterator_from_dataset}
\title{Create an iterator for enumerating the elements of this dataset.}
\usage{
iterator_from_dataset(dataset, features = NULL, response = NULL,
  named_features = TRUE)
}
\arguments{
\item{dataset}{A dataset}

\item{features}{Featured to include. When \code{features} is specified the
iterator value will be either:
\itemize{
\item \code{list(list(feature_name = feature_value, ...), response_value)} when
\code{named_features} is \code{TRUE}; or
\item \code{list(features_array, response_value)} when \code{named_features} is \code{FALSE},
where \code{features_array} is a Rank 2 array of (batch_size, num_features).
}}

\item{response}{Response variable (required from \code{features} is specified).}
}
\value{
An iterator over the elements of this dataset.
}
\description{
The returned iterator will be initialized automatically. The iterator is
"one-shot" (i.e. it performs a single pass over the dataset and cannot be
re-initialized)
}
\details{
Pass the iterator to \code{\link[=next_element]{next_element()}} in order to retreive the value
of the next element, or \code{\link[=next_element_tensor]{next_element_tensor()}} to retreive a tensor that
can be evaluated repeatedly to retreive the value of the next element.
}
