% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iterator_initializer}
\alias{iterator_initializer}
\title{An operation that should be run to initialize this iterator.}
\usage{
iterator_initializer(iterator)
}
\arguments{
\item{iterator}{An iterator}
}
\description{
An operation that should be run to initialize this iterator.
}
\seealso{
Other iterator functions: \code{\link{iterator_get_next}},
  \code{\link{iterator_make_initializer}},
  \code{\link{iterator_string_handle}},
  \code{\link{make-iterator}}
}
