% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iterator_make_initializer}
\alias{iterator_make_initializer}
\title{Create an operation that can be run to initialize this iterator}
\usage{
iterator_make_initializer(iterator, dataset, name = NULL)
}
\arguments{
\item{iterator}{An iterator}

\item{dataset}{A dataset}

\item{name}{(Optional) A name for the created operation.}
}
\value{
A tf$Operation that can be run to initialize this iterator on the
given dataset.
}
\description{
Create an operation that can be run to initialize this iterator
}
\seealso{
Other iterator functions: \code{\link{iterator_get_next}},
  \code{\link{iterator_initializer}},
  \code{\link{iterator_string_handle}},
  \code{\link{make-iterator}}
}
