% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{iterator_next}
\alias{iterator_next}
\title{Get the next values from an iterator}
\usage{
iterator_next(iter, session, completed = NULL)
}
\arguments{
\item{iter}{Dataset iterator}

\item{session}{TensorFlow session to evaluate iterator within}

\item{completed}{Sentinel value to return from \code{iterator_next()} if the
iteration completes (defaults to \code{NULL} but can be any value you specify).}
}
\value{
Nested structure of tensors containing the next element.
}
\description{
Returns the evaluation of the next tensor(s) available from the iterator or
the sentinel \code{completed} value if there are no more tensors to iterate.
}
\seealso{
Other dataset iterators: \code{\link{initializable_iterator}},
  \code{\link{one_shot_iterator}}
}
