% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iterator_string_handle}
\alias{iterator_string_handle}
\title{String-valued tensor that represents this iterator}
\usage{
iterator_string_handle(iterator, name = NULL)
}
\arguments{
\item{iterator}{An iterator}

\item{name}{(Optional) A name for the created operation.}
}
\value{
Scalar tensor of type string
}
\description{
String-valued tensor that represents this iterator
}
\seealso{
Other iterator functions: 
\code{\link{iterator_get_next}()},
\code{\link{iterator_initializer}()},
\code{\link{iterator_make_initializer}()},
\code{\link{make-iterator}}
}
\concept{iterator functions}
