% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{make_dataset_iterator}
\alias{make_dataset_iterator}
\title{Creates an iterator that works with \code{flowery}.}
\usage{
make_dataset_iterator(x)
}
\arguments{
\item{x}{a tf dataset object.}
}
\description{
Creates an iterator that works with \code{flowery}.
}
