% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{out_of_range_error}
\alias{out_of_range_error}
\title{Check if the last TensorFlow error was OutOfRangeError}
\usage{
out_of_range_error()
}
\value{
\code{TRUE} if the last error was OutOfRangeError
}
\description{
Used to detect end of iteration when evaluating tensors returned by
\code{\link[=batches_from_dataset]{batches_from_dataset()}}.
}
\keyword{internal}
