% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{out_of_range_handler}
\alias{out_of_range_handler}
\title{Handle out of range errors on dataset iterators}
\usage{
out_of_range_handler(e)
}
\arguments{
\item{e}{R error object}
}
\description{
Handle out of range errors on dataset iterators
}
\details{
When a dataset iterator reaches the end, an out of range runtime error
will occur. You can catch and ignore the error when it occurs by using
\code{out_of_range_handler} as the \code{error} argument to \code{tryCatch()}.
}
\examples{
\dontrun{
library(tfdatasets)
dataset <- text_line_dataset("mtcars.csv", record_spec = mtcars_spec) \%>\%
  dataset_prepare(x = c(mpg, disp), y = cyl) \%>\%
  dataset_batch(128) \%>\%
  dataset_repeat(10)

batch <- next_batch(dataset)

tryCatch({
  while(TRUE) {
    batch <- sess$run(next_batch)
    # use batch$x and batch$y tensors
  }
}, error = out_of_range_handler)
}

}
