% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_properties.R
\name{output_types}
\alias{output_types}
\alias{output_shapes}
\title{Dataset output types and shapes}
\usage{
output_types(dataset)

output_shapes(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
\code{output_types()} returns the type of each component of an element of
this dataset; \code{output_shapes()} returns the shape of each component of an
element of this dataset.
}
\description{
Dataset output types and shapes
}
