% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{random_integer_dataset}
\alias{random_integer_dataset}
\title{Creates a \code{Dataset} of pseudorandom values}
\usage{
random_integer_dataset(seed = NULL)
}
\arguments{
\item{seed}{(Optional) If specified, the dataset produces a deterministic
sequence of values.}
}
\description{
Creates a \code{Dataset} of pseudorandom values
}
\details{
The dataset generates a sequence of uniformly distributed integer values (dtype int64).
}
