% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_dataset.R
\name{range_dataset}
\alias{range_dataset}
\title{Creates a dataset of a step-separated range of values.}
\usage{
range_dataset(from = 0, to = 0, by = 1, ..., dtype = tf$int64)
}
\arguments{
\item{from}{Range start}

\item{to}{Range end (exclusive)}

\item{by}{Increment of the sequence}

\item{...}{ignored}

\item{dtype}{Output dtype. (Optional, default: \code{tf$int64}).}
}
\description{
Creates a dataset of a step-separated range of values.
}
