% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_spec.R
\name{record_spec}
\alias{record_spec}
\title{Specification for reading a record}
\usage{
record_spec(names, types = NULL, defaults = NULL)
}
\arguments{
\item{names}{Character vector with column names.}

\item{types}{Character vector with olumn types. If \code{NULL} and \code{defaults} is
specified then types will be imputed from the defaults.

Types can be explicitliy specified in a character vector as "integer",
"double", and "character" (e.g. \code{types = c("double", "double", "integer"}).

Alternatively, you can use a compact string representation where each
character represents one column: c = character, i = integer, d = double
(e.g. \code{types =}ddi`).}

\item{defaults}{List of default values which are used when data is missing
from a record (e.g. \code{list(0, 0, 0L}). If \code{NULL} then defaults will be
automatically provided based on \code{types} (\code{0} for numeric columns and \code{""}
for character columns).}
}
\value{
Record specification to be passed to a dataset read function.
}
\description{
Specification for reading a record
}
