% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{tf}
\alias{shape}
\alias{install_tensorflow}
\alias{fit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{tensorflow}{\code{\link[tensorflow]{install_tensorflow}}, \code{\link[tensorflow]{shape}}, \code{\link[tensorflow]{tf}}}

  \item{tidyselect}{\code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

