% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{scaler}
\alias{scaler}
\title{List of pre-made scalers}
\description{
\itemize{
\item \link{scaler_standard}: mean and standard deviation normalizer.
\item \link{scaler_min_max}: min max normalizer
}
}
\seealso{
\link{step_numeric_column}
}
