% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{scaler_min_max}
\alias{scaler_min_max}
\title{Creates an instance of a min max scaler}
\usage{
scaler_min_max()
}
\description{
This scaler will learn the min and max of the numeric variable
and use this to create a \code{normalizer_fn}.
}
\seealso{
\link{scaler} to a complete list of normalizers

Other scaler: 
\code{\link{scaler_standard}()}
}
\concept{scaler}
