% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{scaler_standard}
\alias{scaler_standard}
\title{Creates an instance of a standard scaler}
\usage{
scaler_standard()
}
\description{
This scaler will learn the mean and the standard deviation
and use this to create a \code{normalizer_fn}.
}
\seealso{
\link{scaler} to a complete list of normalizers

Other scaler: 
\code{\link{scaler_min_max}()}
}
\concept{scaler}
