% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors_dataset.R
\name{sparse_tensor_slices_dataset}
\alias{sparse_tensor_slices_dataset}
\title{Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.}
\usage{
sparse_tensor_slices_dataset(sparse_tensor, shuffle = NULL,
  batch_size = NULL, repeated = NULL)
}
\arguments{
\item{sparse_tensor}{A \code{tf$SparseTensor}.}

\item{shuffle}{Randomly sample elements from this dataset (the specified value is
used as the buffer size for shuffling).}

\item{batch_size}{Combine consecutive elements of this dataset into batches
of the specified \code{batch_size}.}

\item{repeated}{An integer value representing the number of times
the elements of this dataset should be repeated.}
}
\value{
A dataset of rank-(N-1) sparse tensors.
}
\description{
Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.
}
\seealso{
Other tensor datasets: \code{\link{tensor_slices_dataset}},
  \code{\link{tensors_dataset}}
}
