% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dataset.R
\name{sql_dataset}
\alias{sql_dataset}
\alias{sqlite_dataset}
\title{A dataset consisting of the results from a SQL query}
\usage{
sql_dataset(driver_name, data_source_name, query, output_types)

sqlite_dataset(filename, query, output_types)
}
\arguments{
\item{driver_name}{String containing the database type. Currently, the only
supported value is 'sqlite'.}

\item{data_source_name}{String containing a connection string to connect to
the database.}

\item{query}{String containing the SQL query to execute.}

\item{output_types}{List of \code{tf$DType`` objects (e.g.}tf$int32\code{,}tf$float32`) representing the types of the columns returned by query.}

\item{filename}{Filename for the database}
}
\value{
A dataset
}
\description{
A dataset consisting of the results from a SQL query
}
