% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\docType{data}
\name{StandardScaler}
\alias{StandardScaler}
\title{http://notmatthancock.github.io/2017/03/23/simple-batch-stat-updates.html
batch updates for mean and variance.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StandardScaler
}
\description{
http://notmatthancock.github.io/2017/03/23/simple-batch-stat-updates.html
batch updates for mean and variance.
}
\keyword{datasets}
