% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{step_image_embedding_column}
\alias{step_image_embedding_column}
\title{Creates Image embeddings columns}
\usage{
step_image_embedding_column(spec, ..., module_spec)
}
\description{
Use this step to create image embeddings columns from image columns.
}
