% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{step_text_embedding_column}
\alias{step_text_embedding_column}
\title{Creates text embeddings columns}
\usage{
step_text_embedding_column(spec, ..., module_spec, trainable = FALSE)
}
\arguments{
\item{spec}{A feature specification created with \code{\link[=feature_spec]{feature_spec()}}.}

\item{...}{Comma separated list of variable names to apply the step. \link{selectors} can also be used.}

\item{module_spec}{A string handle or a _ModuleSpec identifying the module.}

\item{trainable}{Whether or not the Module is trainable. \code{FALSE} by default,
meaning the pre-trained weights are frozen. This is different from
the ordinary.}
}
\description{
Use this step to create text embeddings columns from string columns.
}
