% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{steps}
\alias{steps}
\title{Steps for feature columns specification.}
\description{
List of steps that can be used to specify columns in the \code{feature_spec} interface.
}
\section{Steps}{

\itemize{
\item \code{\link[=step_numeric_column]{step_numeric_column()}} to define numeric columns.
\item \code{\link[=step_categorical_column_with_vocabulary_list]{step_categorical_column_with_vocabulary_list()}} to define categorical columns.
\item \code{\link[=step_categorical_column_with_hash_bucket]{step_categorical_column_with_hash_bucket()}}
\item \code{\link[=step_categorical_column_with_identity]{step_categorical_column_with_identity()}}
\item \code{\link[=step_categorical_column_with_vocabulary_file]{step_categorical_column_with_vocabulary_file()}}
\item \code{\link[=step_indicator_column]{step_indicator_column()}}
\item \code{\link[=step_embedding_column]{step_embedding_column()}}
\item \code{\link[=step_bucketized_column]{step_bucketized_column()}}
\item \code{\link[=step_crossed_column]{step_crossed_column()}}
\item \code{\link[=step_shared_embeddings_column]{step_shared_embeddings_column()}}
}
}

\seealso{
\itemize{
\item \link{selectors} for a list of selectors that can be used to specify variables.
}

Other Feature Spec Functions: \code{\link{dataset_use_spec}},
  \code{\link{feature_spec}}, \code{\link{fit.FeatureSpec}}
}
\concept{Feature Spec Functions}
