% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{with_dataset_iterator}
\alias{with_dataset_iterator}
\title{Execute code that checks for end of dataset iteration}
\usage{
with_dataset_iterator(expr)
}
\arguments{
\item{expr}{Expression to execute}
}
\description{
Execute code that checks for end of dataset iteration
}
\details{
When a dataset iterator reaches the end, an out of range runtime error
will occur. You can catch and ignore the error when it occurs by wrapping
your iteration code in the \code{with_dataset_iterator()} (see the example
below for an illustration).
}
\examples{
\dontrun{
library(tfdatasets)
dataset <- csv_dataset("training.csv") \%>\%
  dataset_batch(128) \%>\%
  dataset_repeat(10)

batch <- batch_from_dataset(dataset, features = c(mpg, disp), response = cyl)
with_dataset_iterator({
  while(TRUE) {
    # use batch$x and batch$y tensors
  }
})
}

}
