% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip-dataset.R
\name{zip_datasets}
\alias{zip_datasets}
\title{Creates a dataset by zipping together the given datasets.}
\usage{
zip_datasets(datasets)
}
\arguments{
\item{datasets}{A nested structure (e.g. list or list of lists) of datasets.}
}
\value{
A dataset
}
\description{
Merges datasets together into pairs or tuples that contain an element from
each dataset.
}
