% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_line_dataset.R
\name{csv_dataset}
\alias{csv_dataset}
\title{Create a dataset from a text file with comma separated values}
\usage{
csv_dataset(filenames, compression_type = NULL, col_names = NULL,
  record_defaults = NULL, field_delim = ",", skip = 0,
  num_parallel_calls = NULL)
}
\arguments{
\item{filenames}{Characater vector containing one or more filenames or
filename glob patterns (e.g. "train.csv", "\emph{.csv", "}-train.csv", etc.)}

\item{compression_type}{A string, one of: \code{"auto"} (determine based on file
extension), \code{""} (no compression), \code{"ZLIB"}, or \code{"GZIP"}. For
\code{"auto"}, GZIP will be automatically selected if any of
the \code{filenames} have a .gz extension and ZLIB will be automatically
selected if any of the \code{filenames} have a .zlib extension (otherwise
no compression will be used).}

\item{col_names}{Character vector with column names (or \code{NULL} to automatically
detect the column names from the first row of the input file).

If \code{col_names} is a character vector, the values will be used as the names
of the columns, and the first row of the input will be read into the first
row of the datset. Note that if the underlying CSV file also includes
column names in it's first row, this row should be skipped explicitly with
\code{\link[=dataset_skip]{dataset_skip()}}.

If \code{NULL}, the first row of the input will be used as the column names, and
will not be included in dataset.}

\item{record_defaults}{List of default values for records. Default values
must be of type integer, numeric, or character. This is used both to indicate the
type of each field as well as to provide defaults for missing values.
If you want all fields to default to a certain type (e.g. numeric) then
you can pass a single value or type specifier (e.g. \code{record_defaults = 0} or
\code{record_defaults = "numeric"}).}

\item{field_delim}{An optional string. Defaults to ",". char delimiter to
separate fields in a record.}

\item{skip}{Number of lines to skip before reading data. Note that if
\code{col_names} is explicitly provided and there are column names witin the CSV
file then \code{skip} should be set to 1 to ensure that the column names are
bypassed.}

\item{num_parallel_calls}{(Optional) An integer, representing the
number of elements to process in parallel If not specified, elements will
be processed sequentially.}
}
\description{
Create a dataset from a text file with comma separated values
}
\note{
The \code{\link[=csv_dataset]{csv_dataset()}} function is a convenience wrapper for the
\code{\link[=text_line_dataset]{text_line_dataset()}} and \code{\link[=dataset_decode_csv]{dataset_decode_csv()}} functions.
}
