% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_line_dataset.R
\name{dataset_decode_csv}
\alias{dataset_decode_csv}
\title{Transform a dataset with CSV text lines into a dataset with named columns}
\usage{
dataset_decode_csv(dataset, col_names = NULL, record_defaults = NULL,
  field_delim = ",", num_parallel_calls = NULL)
}
\arguments{
\item{dataset}{Dataset with CSV text lines}

\item{col_names}{Character vector with column names (or \code{NULL} to automatically
detect the column names from the first row of the input file).

If \code{col_names} is a character vector, the values will be used as the names
of the columns, and the first row of the input will be read into the first
row of the datset. Note that if the underlying CSV file also includes
column names in it's first row, this row should be skipped explicitly with
\code{\link[=dataset_skip]{dataset_skip()}}.

If \code{NULL}, the first row of the input will be used as the column names, and
will not be included in dataset.}

\item{record_defaults}{List of default values for records. Default values
must be of type integer, numeric, or character. This is used both to indicate the
type of each field as well as to provide defaults for missing values.
If you want all fields to default to a certain type (e.g. numeric) then
you can pass \code{record_defaults = "numeric"}.}

\item{field_delim}{An optional string. Defaults to ",". char delimiter to
separate fields in a record.}

\item{num_parallel_calls}{(Optional) An integer, representing the
number of elements to process in parallel If not specified, elements will
be processed sequentially.}
}
\description{
Transform a dataset with CSV text lines into a dataset with named columns
}
\note{
The \code{\link[=csv_dataset]{csv_dataset()}} function is a convenience wrapper for the
\code{\link[=text_line_dataset]{text_line_dataset()}} and \code{\link[=dataset_decode_csv]{dataset_decode_csv()}} functions.
}
\seealso{
Other dataset methods: \code{\link{dataset_batch}},
  \code{\link{dataset_cache}},
  \code{\link{dataset_concatenate}},
  \code{\link{dataset_map}}, \code{\link{dataset_repeat}},
  \code{\link{dataset_shuffle}},
  \code{\link{dataset_skip}}, \code{\link{dataset_take}}
}
