% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_repeat}
\alias{dataset_repeat}
\title{Repeats a dataset count times.}
\usage{
dataset_repeat(dataset, count = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{count}{(Optional.) An integer value representing the number of times
the elements of this dataset should be repeated. The default behavior (if
\code{count} is \code{NULL} or \code{-1}) is for the elements to be repeated indefinitely.}
}
\value{
A dataset
}
\description{
Repeats a dataset count times.
}
\seealso{
Other dataset methods: 
\code{\link{dataset_batch}()},
\code{\link{dataset_cache}()},
\code{\link{dataset_collect}()},
\code{\link{dataset_concatenate}()},
\code{\link{dataset_decode_delim}()},
\code{\link{dataset_filter}()},
\code{\link{dataset_interleave}()},
\code{\link{dataset_map_and_batch}()},
\code{\link{dataset_map}()},
\code{\link{dataset_padded_batch}()},
\code{\link{dataset_prefetch_to_device}()},
\code{\link{dataset_prefetch}()},
\code{\link{dataset_reduce}()},
\code{\link{dataset_shuffle_and_repeat}()},
\code{\link{dataset_shuffle}()},
\code{\link{dataset_skip}()},
\code{\link{dataset_take_while}()},
\code{\link{dataset_take}()},
\code{\link{dataset_window}()}
}
\concept{dataset methods}
