% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_unbatch}
\alias{dataset_unbatch}
\title{Unbatch a dataset}
\usage{
dataset_unbatch(dataset, name = NULL)
}
\arguments{
\item{dataset}{A dataset}

\item{name}{(Optional.) A name for the tf.data operation.}
}
\description{
Splits elements of a dataset into multiple elements.
}
