% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list_dataset.R
\name{file_list_dataset}
\alias{file_list_dataset}
\title{A dataset of all files matching a pattern}
\usage{
file_list_dataset(file_pattern)
}
\arguments{
\item{file_pattern}{A string, representing the filename pattern that will be
matched.}
}
\value{
A dataset of string correponding to file names
}
\description{
A dataset of all files matching a pattern
}
\details{
For example, if we had the following files on our filesystem:
\itemize{
\item /path/to/dir/a.txt
\item /path/to/dir/b.csv
\item /path/to/dir/c.csv
}

If we pass "/path/to/dir/*.csv" as the \code{file_pattern}, the dataset would produce:
\itemize{
\item /path/to/dir/b.csv
\item /path/to/dir/c.csv
}
}
