% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{for_each_batch}
\alias{for_each_batch}
\title{Iterate over batch tensors}
\usage{
for_each_batch(batches, f)
}
\arguments{
\item{batches}{Tensor(s) from \code{\link[=batch_from_dataset]{batch_from_dataset()}}}

\item{f}{Function with signature \code{function(batch)} to
call for each training batch .}
}
\description{
Iterate over batch tensors
}
\note{
The provided function will be called until all
available training batches have been yielded. You
can return \code{FALSE} from the function to stop iteration
early.
}
\examples{
\dontrun{
library(tfdatasets)
dataset <- csv_dataset("training.csv") \%>\%
  dataset_batch(128) \%>\%
  dataset_repeat(10)
batch <- batch_from_dataset(dataset)
for_each_batch(batch, function(batch) {
  # use batch$x and batch$y tensors
  # (return FALSE to stop iteration early)
}
}
}
