% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iterator_get_next}
\alias{iterator_get_next}
\title{Get next element from iterator}
\usage{
iterator_get_next(iterator, name = NULL)
}
\arguments{
\item{iterator}{An iterator}

\item{name}{(Optional) A name for the created operation.}
}
\value{
A nested list of tensors
}
\description{
Returns a nested list of tensors that when evaluated will yield
the next element(s) in the dataset.
}
\seealso{
Other iterator functions: 
\code{\link{iterator_initializer}()},
\code{\link{iterator_make_initializer}()},
\code{\link{iterator_string_handle}()},
\code{\link{make-iterator}}
}
\concept{iterator functions}
