% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_iterators.R
\name{one_shot_iterator}
\alias{one_shot_iterator}
\title{Creates an iterator for enumerating the elements of this dataset.}
\usage{
one_shot_iterator(dataset)
}
\value{
An iterator over the elements of this dataset
}
\description{
The returned iterator will be initialized automatically. A "one-shot"
iterator does not currently support re-initialization.
}
