% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_properties.R
\name{output_types}
\alias{output_types}
\alias{output_shapes}
\title{Output types and shapes}
\usage{
output_types(object)

output_shapes(object)
}
\arguments{
\item{object}{A dataset or iterator}
}
\value{
\code{output_types()} returns the type of each component of an element of
this object; \code{output_shapes()} returns the shape of each component of an
element of this object
}
\description{
Output types and shapes
}
