% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\docType{data}
\name{selectors}
\alias{selectors}
\alias{cur_info_env}
\title{Selectors}
\format{An object of class \code{environment} of length 0.}
\usage{
cur_info_env
}
\description{
List of selectors that can be used to specify variables inside
steps.
}
\section{Selectors}{

\itemize{
\item \code{\link[=has_type]{has_type()}}
\item \code{\link[=all_numeric]{all_numeric()}}
\item \code{\link[=all_nominal]{all_nominal()}}
\item \code{\link[=starts_with]{starts_with()}}
\item \code{\link[=ends_with]{ends_with()}}
\item \code{\link[=one_of]{one_of()}}
\item \code{\link[=matches]{matches()}}
\item \code{\link[=contains]{contains()}}
\item \code{\link[=everything]{everything()}}
}
}

\keyword{datasets}
