% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors_dataset.R
\name{sparse_tensor_slices_dataset}
\alias{sparse_tensor_slices_dataset}
\title{Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.}
\usage{
sparse_tensor_slices_dataset(sparse_tensor)
}
\arguments{
\item{sparse_tensor}{A \code{tf$SparseTensor}.}
}
\value{
A dataset of rank-(N-1) sparse tensors.
}
\description{
Splits each rank-N \code{tf$SparseTensor} in this dataset row-wise.
}
\seealso{
Other tensor datasets: 
\code{\link{tensor_slices_dataset}()},
\code{\link{tensors_dataset}()}
}
\concept{tensor datasets}
