% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{StandardScaler}
\alias{StandardScaler}
\title{http://notmatthancock.github.io/2017/03/23/simple-batch-stat-updates.html
batch updates for mean and variance.}
\description{
http://notmatthancock.github.io/2017/03/23/simple-batch-stat-updates.html
batch updates for mean and variance.

http://notmatthancock.github.io/2017/03/23/simple-batch-stat-updates.html
batch updates for mean and variance.
}
\section{Super class}{
\code{\link[tfdatasets:Scaler]{tfdatasets::Scaler}} -> \code{StandardScaler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit_batch}{\code{StandardScaler$fit_batch()}}
\item \href{#method-fit_resume}{\code{StandardScaler$fit_resume()}}
\item \href{#method-fun}{\code{StandardScaler$fun()}}
\item \href{#method-clone}{\code{StandardScaler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_batch"></a>}}
\subsection{Method \code{fit_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$fit_batch(batch)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_resume"></a>}}
\subsection{Method \code{fit_resume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$fit_resume()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fun"></a>}}
\subsection{Method \code{fun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$fun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
