% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_spec.R
\name{step_image_embedding_column}
\alias{step_image_embedding_column}
\title{Creates Image embeddings columns}
\usage{
step_image_embedding_column(spec, ..., module_spec)
}
\arguments{
\item{spec}{A feature specification created with \code{\link[=feature_spec]{feature_spec()}}.}

\item{...}{Comma separated list of variable names to apply the step. \link{selectors} can also be used.}

\item{module_spec}{A string handle or a _ModuleSpec identifying the module.}
}
\description{
Use this step to create image embeddings columns from image columns.
}
