% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors_dataset.R
\name{tensor_slices_dataset}
\alias{tensor_slices_dataset}
\title{Creates a dataset whose elements are slices of the given tensors.}
\usage{
tensor_slices_dataset(tensors)
}
\arguments{
\item{tensors}{A nested structure of tensors, each having the same size in
the first dimension.}
}
\value{
A dataset.
}
\description{
Creates a dataset whose elements are slices of the given tensors.
}
\seealso{
Other tensor datasets: 
\code{\link{sparse_tensor_slices_dataset}()},
\code{\link{tensors_dataset}()}
}
\concept{tensor datasets}
