% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors_dataset.R
\name{tensors_dataset}
\alias{tensors_dataset}
\title{Creates a dataset with a single element, comprising the given tensors.}
\usage{
tensors_dataset(tensors)
}
\arguments{
\item{tensors}{A nested structure of tensors.}
}
\value{
A dataset.
}
\description{
Creates a dataset with a single element, comprising the given tensors.
}
\seealso{
Other tensor datasets: 
\code{\link{sparse_tensor_slices_dataset}()},
\code{\link{tensor_slices_dataset}()}
}
\concept{tensor datasets}
