% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_line_dataset.R
\name{text_line_dataset}
\alias{text_line_dataset}
\title{Create a dataset comprising lines from one or more text files.}
\usage{
text_line_dataset(filenames, compression_type = NULL)
}
\arguments{
\item{filenames}{Characater vector containing one or more filenames.}

\item{compression_type}{A string, one of: \code{""} (no compression), \code{"ZLIB"}, or \code{"GZIP"}.}
}
\value{
A dataset
}
\description{
Create a dataset comprising lines from one or more text files.
}
