% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrecord_dataset.R
\name{tfrecord_dataset}
\alias{tfrecord_dataset}
\title{A dataset comprising records from one or more TFRecord files.}
\usage{
tfrecord_dataset(filenames, compression_type = NULL)
}
\arguments{
\item{filenames}{String(s) specifying one or more filenames}

\item{compression_type}{A string, one of: \code{NULL} (no compression), \code{"ZLIB"}, or
\code{"GZIP"}.}
}
\description{
A dataset comprising records from one or more TFRecord files.
}
